scripts/requirements.txt
# Environment variables
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Data generation
pandas>=2.0.0
faker>=20.0.0
